+++ src/device/jupiter.ts
@@ -11,6 +11,7 @@ import {
   JupiterBMSInfo,
   JupiterMPPTPVInfo,
   isValidJupiterWorkingMode,
+  isValidJupiterRechargeMode,
   WeekdaySet,
   JupiterTimePeriod,
 } from '../types';
@@ -45,6 +46,7 @@ enum CommandType {
   SET_DEVICE_TIME = 4,
   SET_TIME_PERIOD = 3,
   SET_WORKING_MODE = 2,
+  SET_METER_TYPE = 18,
   SURPLUS_FEED_IN = 13,
   GET_BMS_INFO = 14, // -> inv:g_state=1,w_state1=1,w_state2=1,i_err=0,i_war=0,g_vol=2399,g_cur=0,g_pf=0,g_fre=5000,b_vol=526,g_power=119,i_temp=31,mppt:m_state=244,m_err=0,m_temp=30,m_war=0,pv1=350|37|1304,pv2=349|39|1372,pv3=378|18|712,pv4=365|32|1180,b_vol=525,b_cur=85,base_v=221,pe_v=165,fail_t=0,bms:c_vol=571,c_cur=500,d_cur=500,soc=33,soh=100,b_cap=5120,b_vol=5252,b_cur=63,b_temp=250,b_err=0,b_war=0,b_err2=0,b_war2=0,c_flag=192,s_flag=0,b_num=1,vol0=3280,vol1=3281,vol2=3283,vol3=3283,vol4=3283,vol5=3283,vol6=3280,vol7=3284,vol8=3283,vol9=3284,vol10=3282,vol11=3286,vol12=3277,vol13=3286,vol14=3283,vol15=3284,b_temp0=14,b_temp1=15,b_temp2=15,b_temp3=16,env_t=27,mos_t=20,lck=0
   DEPTH_OF_DISCHARGE = 56,
@@ -428,12 +430,28 @@ function registerRuntimeInfoMessage(message: BuildMessageFn) {
         name: 'Phase Type',
       }),
     );
-    field({ key: 'dchrg', path: ['rechargeMode'], transform: number() });
+    field({
+      key: 'dchrg',
+      path: ['rechargeMode'],
+      transform: map(
+        {
+          '0': 'singlePhase',
+          '1': 'threePhase',
+        },
+        'singlePhase',
+      ),
+    });
     advertise(
       ['rechargeMode'],
-      sensorComponent<number>({
+      selectComponent<NonNullable<JupiterDeviceData['rechargeMode']>>({
         id: 'recharge_mode',
         name: 'Recharge Mode',
+        icon: 'mdi:flash',
+        command: 'recharge-mode',
+        valueMappings: {
+          singlePhase: 'Single Phase',
+          threePhase: 'Three Phase',
+        },
       }),
     );
     field({ key: 'dev_n', path: ['deviceVersion'], transform: number() });
@@ -598,6 +616,26 @@ function registerRuntimeInfoMessage(message: BuildMessageFn) {
           params.nl = 1;
         }
         publishCallback(processCommand(CommandType.SET_WORKING_MODE, params));
+      },
+    });
+
+    // Recharge mode (single-/three-phase)
+    command('recharge-mode', {
+      handler: ({ message, publishCallback, updateDeviceState }) => {
+        if (!isValidJupiterRechargeMode(message)) {
+          logger.warn('Invalid recharge mode value:', message);
+          return;
+        }
+
+        updateDeviceState(() => ({
+          rechargeMode: message,
+        }));
+
+        publishCallback(
+          processCommand(CommandType.SET_METER_TYPE, {
+            dchrg: message === 'threePhase' ? 1 : 0,
+          }),
+        );
       },
     });